[#575] Increase unit test coverage for bot store middleware

- Created 100% unit test coverage for makeBotGreetUser
- Refactored makeBotGreetUser
